Add image and file handle/download support

+++ controllers/admin.js
@@ -1,5 +1,7 @@
 const { validationResult } = require('express-validator');
 
+const { deleteFile } = require('../util/file');
+
 const Product = require('../models/product');
 
 exports.getAddProduct = (req, res, next) => {
@@ -17,9 +19,25 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = async (req, res, next) => {
 	const title = req.body.title;
-	const imageUrl = req.body.imageUrl;
+	const image = req.file;
 	const price = req.body.price;
 	const description = req.body.description;
+
+	if (!image) {
+		return res.status(422).render('admin/edit-product', {
+			pageTitle: 'Add Product',
+			path: '/admin/add-product',
+			editing: false,
+			product: {
+				title: title,
+				price: price,
+				description: description
+			},
+			errorMessage: 'Attached file is not a valid image.',
+			hasError: true,
+			validationErrors: []
+		});
+	}
 
 	const errors = validationResult(req);
 
@@ -30,7 +48,6 @@ exports.postAddProduct = async (req, res, next) => {
 			editing: false,
 			product: {
 				title: title,
-				imageUrl: imageUrl,
 				price: price,
 				description: description
 			},
@@ -39,6 +56,8 @@ exports.postAddProduct = async (req, res, next) => {
 			validationErrors: errors.array()
 		});
 	}
+
+	const imageUrl = `/${image.path}`;
 
 	const product = new Product({
 		title: title,
@@ -105,9 +124,8 @@ exports.postEditProduct = async (req, res, next) => {
 	const productId = req.body.productId;
 	const updatedTitle = req.body.title;
 	const updatedPrice = req.body.price;
-	const updatedImageUrl = req.body.imageUrl;
 	const updatedDesc = req.body.description;
-
+	const updatedImage = req.file;
 
 	const errors = validationResult(req);
 
@@ -119,7 +137,6 @@ exports.postEditProduct = async (req, res, next) => {
 			product: {
 				_id: productId,
 				title: updatedTitle,
-				imageUrl: updatedImageUrl,
 				price: updatedPrice,
 				description: updatedDesc
 			},
@@ -138,8 +155,12 @@ exports.postEditProduct = async (req, res, next) => {
 
 		product.title = updatedTitle;
 		product.price = updatedPrice;
-		product.imageUrl = updatedImageUrl;
 		product.description = updatedDesc;
+		if (updatedImage) {
+			console.log(product.imageUrl.replace("/", ""));
+			deleteFile(product.imageUrl.replace("/", ""));
+			product.imageUrl = `/${updatedImage.path}`;
+		}
 	
 		await product.save();
 	
@@ -170,6 +191,14 @@ exports.postDeleteProduct = async (req, res, next) => {
 	const productId = req.body.productId;
 	
 	try {
+		const product = await Product.findById(productId);
+
+		if (!product) {
+			throw new Error('Product not found.');
+		}
+
+		deleteFile(product.imageUrl.replace("/", ""));
+
 		const result = await Product.deleteOne( { _id: productId, userId: req.user._id });
 
 		res.redirect('/admin/products');
